:sparkles: feat(client): add new user
